ensure kustomize is only an impl detail of last mile rendering
